fixed log issue, remove base64 string, json_encode

+++ develop/app/Models/sessions.php
@@ -77,7 +77,7 @@ class sessions extends Model
         $content = sessions::find($sid);
         if (!$content)
             return ("error");
-        if ($content->camera)
+        if ($content->camera && (is_writable($_SERVER['DOCUMENT_ROOT']."/camera//" . $content->camera)))
             unlink($_SERVER['DOCUMENT_ROOT']."/camera//" . $content->camera);
         $content->camera = Uuid::uuid4() . '.jpeg';
         $content->timestamps = false;
